tested bnb and link after adding full, optimized, and ablation folder

+++ optimizations/references/test_longTrace/link_test_gas.js
@@ -71,10 +71,13 @@ helper.range(transactionCounts).forEach(l => {
       let mintAmount_1 = helper.random(lowerBound, upperBound+1);
       let mintAmount_2 = helper.random(lowerBound, upperBound+1);
       let text = `mint,constructor,,${totalSupply},${constructorSenderIndex},,false\nmint,mint,instance,accounts[${mintAccountIndex}] ${mintAmount_1},${constructorSenderIndex},,false\nmint,mint,instance,accounts[${mintAccountIndex}] ${mintAmount_2},${constructorSenderIndex},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     }) 
   }
 
@@ -95,10 +98,13 @@ helper.range(transactionCounts).forEach(l => {
       let approveAmount_1 = helper.random(lowerBound, upperBound+1);
       let approveAmount_2 = approveAmount_1 + helper.random(0, upperBound+1);
       let text = `approve,constructor,,${totalSupply},${constructorSenderIndex},,false\napprove,approve,instance,accounts[${approveAccountIndex}] ${approveAmount_1},${constructorSenderIndex},,false\napprove,approve,instance,accounts[${approveAccountIndex}] ${approveAmount_2},${constructorSenderIndex},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     }) 
   }
 
@@ -114,10 +120,13 @@ helper.range(transactionCounts).forEach(l => {
       let burnAmount_1  = helper.random(1, (mintAmount+1)/2);
       let burnAmount_2  = helper.random(1, (mintAmount+1)/2);
       let text = `burn,constructor,,${totalSupply},${constructorSenderIndex},,false\nburn,mint,instance,accounts[${mintAccountIndex}] ${mintAmount},${constructorSenderIndex},,false\nburn,burn,instance,accounts[${mintAccountIndex}] ${burnAmount_1},${constructorSenderIndex},,false\nburn,burn,instance,accounts[${mintAccountIndex}] ${burnAmount_2},${constructorSenderIndex},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     }) 
   }  
 
@@ -142,10 +151,13 @@ helper.range(transactionCounts).forEach(l => {
       let transferFromAmount_1 = helper.random(1, (approveAmount+1)/2);
       let transferFromAmount_2 = helper.random(1, (approveAmount+1)/2);
       let text = `transferFrom,constructor,,${totalSupply},${constructorSenderIndex},,false\ntransferFrom,approve,instance,accounts[${approveAccountIndex}] ${approveAmount},${constructorSenderIndex},,false\ntransferFrom,transferFrom,instance,accounts[${constructorSenderIndex}] accounts[${transferFromToAccountIndex}] ${transferFromAmount_1},${approveAccountIndex},,false\ntransferFrom,transferFrom,instance,accounts[${constructorSenderIndex}] accounts[${transferFromToAccountIndex}] ${transferFromAmount_2},${approveAccountIndex},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     }) 
   }  
 
@@ -168,10 +180,13 @@ helper.range(transactionCounts).forEach(l => {
       let transferAmount_1 = helper.random(1, (totalSupply+1)/2);
       let transferAmount_2 = helper.random(1, (totalSupply+1)/2);
       let text = `transfer,constructor,,${totalSupply},${constructorSenderIndex},,false\ntransfer,transfer,instance,accounts[${transferAccountIndex}] ${transferAmount_1},${constructorSenderIndex},,false\ntransfer,transfer,instance,accounts[${transferAccountIndex}] ${transferAmount_2},${constructorSenderIndex},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     }) 
   }
   if(transactionName == 'increaseApproval') {
@@ -192,10 +207,13 @@ helper.range(transactionCounts).forEach(l => {
       let increaseApprovalAmount_1 = helper.random(lowerBound, upperBound+1);
       let increaseApprovalAmount_2 = helper.random(lowerBound, upperBound+1);
       let text = `increaseApproval,constructor,,${totalSupply},${constructorSenderIndex},,false\nincreaseApproval,approve,instance,accounts[${approveAccountIndex}] ${approveAmount},${constructorSenderIndex},,false\nincreaseApproval,increaseApproval,instance,accounts[${approveAccountIndex}] ${increaseApprovalAmount_1},${constructorSenderIndex},,false\nincreaseApproval,increaseApproval,instance,accounts[${approveAccountIndex}] ${increaseApprovalAmount_2},${constructorSenderIndex},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     }) 
   }
   if(transactionName == 'decreaseApproval') {
@@ -216,10 +234,13 @@ helper.range(transactionCounts).forEach(l => {
       let decreaseApprovalAmount_1 = helper.random(1, (approveAmount+1)/2);
       let decreaseApprovalAmount_2 = helper.random(1, (approveAmount+1)/2);
       let text = `decreaseApproval,constructor,,${totalSupply},${constructorSenderIndex},,false\ndecreaseApproval,approve,instance,accounts[${approveAccountIndex}] ${approveAmount},${constructorSenderIndex},,false\ndecreaseApproval,decreaseApproval,instance,accounts[${approveAccountIndex}] ${decreaseApprovalAmount_1},${constructorSenderIndex},,false\ndecreaseApproval,decreaseApproval,instance,accounts[${approveAccountIndex}] ${decreaseApprovalAmount_2},${constructorSenderIndex},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     }) 
   }
